extract date picker into abstract edit field component
